JAMES-2293 Move Limit to a more generic project
